The value is too large or too small for an Int32.
System.Number.ParseInt32(String s, NumberStyles style, NumberFormatInfo info)
ExCSS.NumberToken.get_IntegerValue()
ExCSS.ValueExtensions.ToInteger(IEnumerable`1 value)
ExCSS.StructValueConverter`1.Convert(IEnumerable`1 value)
ExCSS.OrValueConverter.Convert(IEnumerable`1 value)
ExCSS.OrValueConverter.Convert(IEnumerable`1 value)
ExCSS.OrValueConverter.Convert(IEnumerable`1 value)
ExCSS.Property.TrySetValue(TokenValue newTokenValue)
ExCSS.StylesheetComposer.CreateDeclarationWith(Func`2 createProperty, Token& token)
ExCSS.StylesheetComposer.FillDeclarations(StyleDeclaration style)
ExCSS.StylesheetComposer.CreateStyle(Token current)
ExCSS.StylesheetComposer.CreateRule(Token token)
ExCSS.StylesheetComposer.CreateRules(Stylesheet sheet)
ExCSS.StylesheetParser.Parse(TextSource source)
ExCSS.StylesheetParser.Parse(String content)

+++ src/ExCSS/Extensions/ValueExtensions.cs
@@ -6,7 +6,7 @@ namespace ExCSS
 {
     internal static class ValueExtensions
     {
-        private static bool IsWeight(int value)
+        private static bool IsWeight(long value)
         {
             return value == 100 || value == 200 || value == 300 || value == 400 ||
                    value == 500 || value == 600 || value == 700 || value == 800 ||
@@ -168,7 +168,7 @@ namespace ExCSS
             return element >= 1f ? element : null;
         }
 
-        public static int? ToInteger(this IEnumerable<Token> value)
+        public static long? ToInteger(this IEnumerable<Token> value)
         {
             var element = value.OnlyOrDefault();
 
@@ -180,25 +180,25 @@ namespace ExCSS
             return null;
         }
 
-        public static int? ToNaturalInteger(this IEnumerable<Token> value)
+        public static long? ToNaturalInteger(this IEnumerable<Token> value)
         {
             var element = value.ToInteger();
             return element >= 0 ? element : null;
         }
 
-        public static int? ToPositiveInteger(this IEnumerable<Token> value)
+        public static long? ToPositiveInteger(this IEnumerable<Token> value)
         {
             var element = value.ToInteger();
             return element > 0 ? element : null;
         }
 
-        public static int? ToWeightInteger(this IEnumerable<Token> value)
+        public static long? ToWeightInteger(this IEnumerable<Token> value)
         {
             var element = value.ToPositiveInteger();
             return element.HasValue && IsWeight(element.Value) ? element : null;
         }
 
-        public static int? ToBinary(this IEnumerable<Token> value)
+        public static long? ToBinary(this IEnumerable<Token> value)
         {
             var element = value.ToInteger();
             return element.HasValue && (element.Value == 0 || element.Value == 1) ? element : null;
